Move copy components to separate file.
