+++ frontend/src/views/ContactUs.tsx
@@ -1,53 +1,22 @@
 import React, { useState } from 'react';
-import styled from 'styled-components';
-import { DefaultButton } from '../components/core/Button';
 import { LandingPageContainer } from '../components/core/Container';
 import { InlineExternalLink } from '../components/core/Link';
 import { ContactHeaderTitle, ContactHeaderText } from '../components/core/Text';
-
-type CopyIndicator = {
-  copiedEmail: boolean,
-}
-
-const CopyIndicatorContainer = styled.div.attrs((props: CopyIndicator) => ({
-  style: {
-    transform: (!props.copiedEmail) ? 'translateY(-60px)' : null,
-  },
-}))<CopyIndicator>`
-  position: absolute;
-  top: 20px;
-  left: 50%;
-  transition: transform 0.25s;
-`;
-
-const CopyIndicator = styled(DefaultButton)`
-  position: relative;
-  left: -50%;
-  margin: 0 auto;
-  padding: 0.25rem 1rem;
-  color: ${({ theme }) => theme.colors.white};
-  background: ${({ theme }) => theme.colors.gradients.green};
-`;
+import {
+  CopyIndicator,
+  CopyIndicatorContainer,
+  InlineCopyIcon,
+  InlineCopyText,
+} from '../components/special/CopyIndicator';
 
 function ContactUsPage() {
-  const InlineEmailCopyText = styled(ContactHeaderText)`
-    display: inline-block;
-    margin: 0;
-    cursor: pointer;
-    border-bottom: 1px solid ${({ theme }) => theme.colors.text};
-  `;
-
-  const CopyIcon = styled.i`
-    margin-left: 5px;
-  `;
-
   const [copiedEmail, setCopiedEmail] = useState<boolean>(false);
 
   return (
     <>
-      <CopyIndicatorContainer copiedEmail={copiedEmail}>
+      <CopyIndicatorContainer copied={copiedEmail}>
         <CopyIndicator onClick={() => setCopiedEmail(false)}>
-          Email copied to clipboard!&nbsp;&nbsp;✕
+          Email copied!&nbsp;&nbsp;✕
         </CopyIndicator>
       </CopyIndicatorContainer>
       <LandingPageContainer>
@@ -70,15 +39,15 @@ function ContactUsPage() {
         <ContactHeaderText>
           You can contact us at
           {' '}
-          <InlineEmailCopyText
+          <InlineCopyText
             onClick={() => {
               navigator.clipboard.writeText('support@codejoust.co');
               setCopiedEmail(true);
             }}
           >
             support@codejoust.co
-            <CopyIcon className="material-icons">content_copy</CopyIcon>
-          </InlineEmailCopyText>
+            <InlineCopyIcon>content_copy</InlineCopyIcon>
+          </InlineCopyText>
           . Say hello!
         </ContactHeaderText>
       </LandingPageContainer>
